addressing comments

+++ psydac/linalg/kron.py
@@ -261,7 +261,7 @@ class KroneckerStencilMatrix(LinearOperator):
         # by using the `diagonal` method of StencilMatrices
         diag = 1.
         for mat, start, end in zip(self.mats[::-1], V.starts[::-1], V.ends[::-1]):
-            diag = np.array([d*diag for d in mat.diagonal(inverse=inverse, sqrt=sqrt)._data[start:end+1]], dtype='float64')
+            diag = np.array([d*diag for d in mat.diagonal(inverse=inverse, sqrt=sqrt)._data[start:end+1]])
 
         if out is not None:
             np.copyto(diag, out._data)
